Completed Recursive function
I completed the SierpinskiTriangle function and changed the triangle
function for the base case. The recursive case has been filled in as
well. There is a limit variable now and the mouseDragged function has
been filled in too. The draw function has a background to reduce the
smudge with fill and noStroke helping the triangle look more clean.